Changed some variables from uint type (which isn't a standard type) to size_t
git-svn-id: http://euclid.esac.esa.int/svn/EC/SGS/SDC/CH/Projects/Alexandria/branches/Alexandria-2.0@4043 463500fe-d6be-4184-9b20-47717e75ecac

+++ ChMath/src/lib/function/multiplication.cpp
@@ -22,8 +22,8 @@ std::unique_ptr<Function> multiplyPolynomials(const Function& f1, const Function
   std::vector<double> c1 = p1.getCoefficients();
   std::vector<double> c2 = p2.getCoefficients();
   std::vector<double> resultCoef(c1.size() + c2.size() - 1, 0.);
-  for (uint i = 0; i < c1.size(); ++i)
-    for (uint j = 0; j < c2.size(); ++j)
+  for (size_t i = 0; i < c1.size(); ++i)
+    for (size_t j = 0; j < c2.size(); ++j)
       resultCoef[i + j] += c1[i] * c2[j];
   return std::unique_ptr<Polynomial>(new Polynomial{resultCoef});
 }
